move hook to usePrepareSendTransactions.ts
